Move posix file methods to separate file
